change API: ZC_endCmpr() and ZC_endDec() and also enable 'automatic release memory' by ZC_Finalize()

+++ zc/src/ZC_DataProperty_double.c
@@ -335,7 +335,7 @@ ZC_DataProperty* ZC_genProperties_double(char* varName, double *data, size_t num
 
 	if(fftFlag)
 	{
-        int fft_size = pow(2, (int)log2(numOfElem));
+        size_t fft_size = pow(2, (int)log2(numOfElem));
         property->fftCoeff = ZC_computeFFT(data, fft_size, ZC_DOUBLE);
 	}
